Add Docker Compose setup for the site.
This commit sets up a docker-compose.yml for use with Docker Compose that sets
up a basic LAMP stack that we can use for the site. It can be started by running
docker-compose from the root directory. The test html pages have been moved into
a proper directory and other files have been dropped.

When the server starts, files are put in the mysql directory that are owned by
root, and thus cannot be entered into the git repository. I have tried to set up
a .gitignore so that these files aren't bothersome, but this might need to be
properly handled at some point.