Replace event names with names pulled from the database.
There are six events listed in events.html, and four events listed in
the SQL database. This simply replaces the titles of the first four
events to those listed in the database.

This is an incredibly fake "solution", but it fulfills my desire to
fetch events from the database, sort of.